requirement.txt
chromadb>=0.4.22
google-generativeai>=0.3.0
pypdf>=3.17.0
python-docx>=1.0.0
